feat: add CoinCollection mechanic & Reset point for player when falling of the island
WIP: this is my second commit, it's still pretty large because i added some assets from the assets store and more materials. Besides I made a CoinCollection mechanic. When the player runs through a coin, it destroys and dissappears, I also fixed an issue where when the player goes beyond the edge of the island, he doesn't fall infinitely, but instead respawns to the initial spawn position.
In the next WIP I will try make a functional coincollection game, with a score count